Quell HTTParty Warnings When Checking `nil?`
HTTParty now throws a warning when you try to call `HTTParty::Response#nil?`:

```
[DEPRECATION] HTTParty will no longer override `response#nil?`. This functionality will be removed in future versions. Please, add explicit check `response.body.nil? || response.body.empty?`. For more info refer to: https://github.com/jnunemaker/httparty/issues/568
```

To resolve this, `require_attr` has been altered to now perform a
`body.nil?` check when the attribute being checked for nil is an
`HTTParty::Response`. This issue was somewhat difficult to reproduce
in your test suite due to the extensive amount of mocking going on, as
well as the fact that `Gemfile.lock` was committed. This warning wasn't
occurring in tests, so I dove in a bit and learned that none of the
tests actually pass a real `HTTParty::Response` into `require_attr`, so
you'd never see this warning in tests but it definitely happens in the
real world. Even if you had been, the fact that `Gemfile.lock` was
committed prevented HTTParty from being updated locally, so the tests
were not actually running against the version of HTTParty that
applications consuming this library have installed.

To resolve this, I deleted `Gemfile.lock` from the repo and added it to
`.gitignore`. This file is not used when you publish to RubyGems.org, as
RubyGems only looks at the `.gemspec` file to figure out dependencies.
This is also the case for consuming applications. It's a general good
practice to not commit `Gemfile.lock` in RubyGem libraries so that
you're always running against the gems that an application would end up
installing as dependencies of your gem.